+++ lib/src/router/router.dart
@@ -25,8 +25,7 @@ class Palace {
   set notFoundHandler(Handler h) => _notFoundHandler = h;
 
   /// get the not found handler id none was assigned it will return the defaults 404 handler;
-  Handler get notFoundHandler =>
-      _notFoundHandler ?? (Request req, Response res) => res.notFound();
+  Handler get notFoundHandler => _notFoundHandler ?? (Request req, Response res) => res.notFound();
 
   /// the server instance
   HttpServer? _server;
@@ -114,27 +113,27 @@ class Palace {
         guards: guards,
       ));
 
-  // void _bootstrap() {
-  // TODO _bootstrap
-  //   for (final e in _endpoints) {
-  //     if (_endpoints.where((element) {
-  //           if (e.method.toUpperCase() == 'ALL') {
-  //             return e.path == element.path;
-  //           } else {
-  //             return e.path == element.path && e.method.toUpperCase() == element.method.toUpperCase();
-  //           }
-  //         }).length >
-  //         1) {
-  //       throw 'your endpoints have a duplicate try to fix it => $e';
-  //     }
-  //   }
-  // }
+  void _bootstrap() {
+    for (final e in _endpoints) {
+      if (_endpoints.where((element) {
+            if (e.method.toUpperCase() == 'ALL') {
+              return e.path == element.path;
+            } else {
+              return e.path == element.path && e.method.toUpperCase() == element.method.toUpperCase();
+            }
+          }).length >
+          1) {
+        throw 'your endpoints have a duplicate try to fix it => $e';
+      }
+    }
+  }
 
   Future<void> openGates({
     int port = 3000,
     String? ip,
     bool enableLogs = true,
   }) async {
+    _bootstrap();
     ip ??= InternetAddress.anyIPv4.address;
     // TODO :: Set the `enableLogs` to true by default in debug only
     if (enableLogs) {
@@ -158,11 +157,7 @@ class Palace {
     /// * wait for incoming requests
     try {
       /// * look for desired endpoint
-      final endpoint = match(ioReq.method, ioReq.uri.path) ??
-          EndPoint(
-              path: ioReq.uri.path,
-              method: ioReq.method,
-              handler: notFoundHandler);
+      final endpoint = match(ioReq.method, ioReq.uri.path) ?? EndPoint(path: ioReq.uri.path, method: ioReq.method, handler: notFoundHandler);
 
       /// * create Place req form dart io req and the desired endpoint;
       final req = await Request.init(ioReq, endpoint);